% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biogeoregions10km.R
\docType{data}
\name{biogeoregions10km}
\alias{biogeoregions10km}
\title{EEA refecence grid (10 km) of biogeographic regions}
\format{
A data frame of 259033 observations and the following 2 variables:
\describe{
  \item{region}{Region code, character}
  \item{CELLCODE}{10 km cell code, character}
}
}
\source{
https://cdr.eionet.europa.eu/help/habitats_art17/Reporting2025
}
\usage{
data(biogeoregions10km)
}
\description{
This dataset consists of 10 km grid cells intersecting each biogeographic region.
}
\examples{
library(ggplot2)
library(dplyr)
library(sf)
data(eu10km)
data(biogeoregions10km)
bgcells <- unname(unlist(select(subset(biogeoregions10km, region=="BOR"),CELLCODE)))
ggplot(subset(eu10km, CELLCODE \%in\% bgcells))+
  geom_sf(fill=NA)+
  theme_void()
}
\keyword{datasets}
