% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu10km.R
\docType{data}
\name{eu10km}
\alias{eu10km}
\title{EEA reference grids (10 km) of member states}
\format{
A simple feature collection with 150446 observations and 1 field:
\describe{
  \item{CELLCODE}{10 km cell code, character}
}
}
\source{
https://sdi.eea.europa.eu/catalogue/srv/api/records/3c362237-daa4-45e2-8c16-aaadfb1a003b
}
\usage{
data(eu10km)
}
\description{
A simple feature collection containing 10 km EEA reference grid cells covering all member states
}
\examples{
library(ggplot2)
library(dplyr)
library(sf)
data(eu10km)
data(ms10kmcells)
mscells <- unname(unlist(select(subset(ms10kmcells, ms=="EE"),CELLCODE)))
ggplot(subset(eu10km, CELLCODE \%in\% mscells))+
  geom_sf(fill=NA)+
  theme_void()
}
\keyword{datasets}
