% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_to_range.R
\name{features_to_range}
\alias{features_to_range}
\title{Calculate distribution and range from input features}
\usage{
features_to_range(
  features,
  reference_grid,
  gap_distance = 40000,
  exclude = NULL
)
}
\arguments{
\item{features}{Input geometries of habitat area (polygons) or
species occurence (points, polygons).}

\item{reference_grid}{Reference grid geometry as sf object of relevant member
state and biogeographical region. Can be prepared with function
'ms_region_reference_grid'.}

\item{gap_distance}{Gap distance parameter in meters. If not specified,
default 40 km.}

\item{exclude}{Cells to be excluded. If specified.}
}
\value{
Outputs a list with elements named 'distribution' and 'range'.
Both consist of list containing an sfc object, with geometry of
distribution/range, vector of cell codes and area in sq km-s.

#' @examples
library(sf)
library(ggplot2)
# load country borders
data(eesti)
# get MS biogeoregion cells
bor_cells_ee <- ms_region_reference_grid("EE", "BOR", eesti)
set.seed(69)
exa_spe_occ <- st_as_sf(st_sample(eesti, 20))
exa_spe_range <- features_to_range(exa_spe_occ, bor_cells_ee)
ggplot(data = exa_spe_range$range$range)+
  geom_sf()+
  geom_sf(data=eesti, fill=NA)+
  geom_sf(data=exa_spe_occ)+
  theme_void()
}
\description{
`features_to_range` function implements the range calculation
procedure described in guidelines of Article 17 reporting under
habitats directive.
}
\details{
This is a function that calculates distribution and range maps from input
features and reference grid. Outputs a list with elements 'distribution' and
'range'. Both consist of list containing an sfc object, with geometry of
distribution/range, vector of cell codes and area in sq km-s.
}
\references{
DG Environment. 2023. Reporting under Article 17
of the Habitats Directive: Guidelines on concepts and definitions
– Article 17 of Directive 92/43/EEC, Reporting period 2019-2024.
Brussels. Pp 104
}
