% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapdistance.R
\docType{data}
\name{gapdistance}
\alias{gapdistance}
\title{Default gap distance for species and habitats}
\format{
A tibble with 1638 row of the following 4 variables:
\describe{
  \item{type}{'species' or 'habitats'}
  \item{code}{Species or habitat code, character}
  \item{group}{Habitat group or species taxonomic group, character}
  \item{gap}{Gap distance in meters}
}
}
\source{
https://cdr.eionet.europa.eu/help/habitats_art17/Reporting2025
}
\usage{
data(gapdistance)
}
\description{
Default gap distance estimates for species and habitats as recommended in guidelines.
}
\keyword{datasets}
